@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 2.8.3)
+cmake_minimum_required(VERSION 3.0.2)
 project(lidar_calibration)
 
 ## Find catkin macros and libraries
@@ -10,7 +10,7 @@ find_package(catkin REQUIRED COMPONENTS
   ceres_catkin
 )
 
-add_compile_options(-std=c++11)
+add_compile_options(-std=c++14)
 
 #SET(CMAKE_BUILD_TYPE RelWithDebInfo)
 
