@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 2.8.3)
+cmake_minimum_required(VERSION 3.0.2)
 project(lidar_calibration_lib)
 
 ## Find catkin macros and libraries
@@ -13,7 +13,7 @@ find_package(catkin REQUIRED COMPONENTS
   tf_conversions
 )
 
-add_compile_options(-std=c++11)
+add_compile_options(-std=c++14)
 
 #SET(CMAKE_BUILD_TYPE RelWithDebInfo)
 
